refactor: export unit template spec generator
