+++ packages/vstory/src/index.ts
@@ -7,3 +7,5 @@ export * from './edit';
 export * from './story/utils/vchart-pick';
 export * from './story/utils/chart';
 export * from './story/utils/layout';
+
+export * from './template/unit';
